Fixed some timer related bugs. Have to look at that later. Also, displayed ELO and usernames on frontend.

+++ backend/src/app.js
@@ -72,6 +72,7 @@ wss.on("connection", (ws, req) => {
 
     ws.on("close", () => {
         // Remove the user from the matchmaking queue if they disconnect
+        if (!ws.user) return;
         matchmakingService.removeFromQueue(ws.user);
         matchmakingService.rankedMatchQueue.removePlayer(ws.user);
     });
